LayoutTests/fast/forms/selection-wrongtype-expected.txt

Tests to ensure selectionStart, selectionEnd, and selectionDirection throw exceptions when the input element is not a text field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test.selectionStart is null
PASS test.selectionStart = 0 threw exception InvalidStateError: The input element's type ('button') does not support selection..
PASS test.selectionEnd is null
PASS test.selectionEnd = 0 threw exception InvalidStateError: The input element's type ('button') does not support selection..
PASS test.selectionDirection is null
PASS test.selectionDirection = 0 threw exception InvalidStateError: The input element's type ('button') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function. (In 'test.selectionRange()', 'test.selectionRange' is undefined).
PASS test.selectionStart is null
PASS test.selectionStart = 0 threw exception InvalidStateError: The input element's type ('checkbox') does not support selection..
PASS test.selectionEnd is null
PASS test.selectionEnd = 0 threw exception InvalidStateError: The input element's type ('checkbox') does not support selection..
PASS test.selectionDirection is null
PASS test.selectionDirection = 0 threw exception InvalidStateError: The input element's type ('checkbox') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function. (In 'test.selectionRange()', 'test.selectionRange' is undefined).
PASS test.selectionStart is null
PASS test.selectionStart = 0 threw exception InvalidStateError: The input element's type ('file') does not support selection..
PASS test.selectionEnd is null
PASS test.selectionEnd = 0 threw exception InvalidStateError: The input element's type ('file') does not support selection..
PASS test.selectionDirection is null
PASS test.selectionDirection = 0 threw exception InvalidStateError: The input element's type ('file') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function. (In 'test.selectionRange()', 'test.selectionRange' is undefined).
PASS test.selectionStart is null
PASS test.selectionStart = 0 threw exception InvalidStateError: The input element's type ('hidden') does not support selection..
PASS test.selectionEnd is null
PASS test.selectionEnd = 0 threw exception InvalidStateError: The input element's type ('hidden') does not support selection..
PASS test.selectionDirection is null
PASS test.selectionDirection = 0 threw exception InvalidStateError: The input element's type ('hidden') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function. (In 'test.selectionRange()', 'test.selectionRange' is undefined).
PASS test.selectionStart is null
PASS test.selectionStart = 0 threw exception InvalidStateError: The input element's type ('image') does not support selection..
PASS test.selectionEnd is null
PASS test.selectionEnd = 0 threw exception InvalidStateError: The input element's type ('image') does not support selection..
PASS test.selectionDirection is null
PASS test.selectionDirection = 0 threw exception InvalidStateError: The input element's type ('image') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function. (In 'test.selectionRange()', 'test.selectionRange' is undefined).
PASS test.selectionStart is null
PASS test.selectionStart = 0 threw exception InvalidStateError: The input element's type ('radio') does not support selection..
PASS test.selectionEnd is null
PASS test.selectionEnd = 0 threw exception InvalidStateError: The input element's type ('radio') does not support selection..
PASS test.selectionDirection is null
PASS test.selectionDirection = 0 threw exception InvalidStateError: The input element's type ('radio') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function. (In 'test.selectionRange()', 'test.selectionRange' is undefined).
PASS test.selectionStart is null
PASS test.selectionStart = 0 threw exception InvalidStateError: The input element's type ('range') does not support selection..
PASS test.selectionEnd is null
PASS test.selectionEnd = 0 threw exception InvalidStateError: The input element's type ('range') does not support selection..
PASS test.selectionDirection is null
PASS test.selectionDirection = 0 threw exception InvalidStateError: The input element's type ('range') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function. (In 'test.selectionRange()', 'test.selectionRange' is undefined).
PASS test.selectionStart is null
PASS test.selectionStart = 0 threw exception InvalidStateError: The input element's type ('reset') does not support selection..
PASS test.selectionEnd is null
PASS test.selectionEnd = 0 threw exception InvalidStateError: The input element's type ('reset') does not support selection..
PASS test.selectionDirection is null
PASS test.selectionDirection = 0 threw exception InvalidStateError: The input element's type ('reset') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function. (In 'test.selectionRange()', 'test.selectionRange' is undefined).
PASS test.selectionStart is null
PASS test.selectionStart = 0 threw exception InvalidStateError: The input element's type ('submit') does not support selection..
PASS test.selectionEnd is null
PASS test.selectionEnd = 0 threw exception InvalidStateError: The input element's type ('submit') does not support selection..
PASS test.selectionDirection is null
PASS test.selectionDirection = 0 threw exception InvalidStateError: The input element's type ('submit') does not support selection..
PASS test.selectionRange() threw exception TypeError: test.selectionRange is not a function. (In 'test.selectionRange()', 'test.selectionRange' is undefined).
PASS successfullyParsed is true

TEST COMPLETE
